Use `u32` Instead of `usize` for Indexing in `backtrack_tree`
This allows us to double the tree size while keeping
memory consumption constant.

+++ src/distributed/worker.rs
@@ -36,7 +36,7 @@ impl Worker {
 
     pub fn run(&mut self) -> Result<()> {
         thread_local! {
-            static STACK_BUF: RefCell<MinMaxHeap<map::MismatchSearchStackFrame>> = RefCell::new(MinMaxHeap::with_capacity(map::STACK_LIMIT + 9));
+            static STACK_BUF: RefCell<MinMaxHeap<map::MismatchSearchStackFrame>> = RefCell::new(MinMaxHeap::with_capacity(map::STACK_LIMIT as usize + 9));
             static TREE_BUF: RefCell<Tree<map::EditOperation>> = RefCell::new(Tree::with_capacity(map::STACK_LIMIT + 9));
         }
